+++ install/migrations/update_10.0.16_to_10.0.17.php
@@ -50,8 +50,6 @@ function update10016to10017()
     $DELFROMDISPLAYPREF = [];
     $update_dir = __DIR__ . '/update_10.0.16_to_10.0.17/';
 
-    //TRANS: %s is the number of new version
-    $migration->displayTitle(sprintf(__('Update to %s'), '10.0.17'));
     $migration->setVersion('10.0.17');
 
     $update_scripts = scandir($update_dir);