+++ install/migrations/update_10.0.5_to_10.0.6.php
@@ -50,8 +50,6 @@ function update1005to1006()
     $DELFROMDISPLAYPREF = [];
     $update_dir = __DIR__ . '/update_10.0.5_to_10.0.6/';
 
-    //TRANS: %s is the number of new version
-    $migration->displayTitle(sprintf(__('Update to %s'), '10.0.6'));
     $migration->setVersion('10.0.6');
 
     $update_scripts = scandir($update_dir);