+++ install/migrations/update_10.0.5_to_10.0.6/configs.php
@@ -36,7 +36,6 @@
  * @var \Migration $migration
  */
 
-$migration->displayMessage('Add new configurations / user preferences');
 $migration->addConfig([
     'timeline_action_btn_layout'   => 0,
     'timeline_date_format'   => 0,