+++ install/migrations/update_10.0.8_to_10.0.9.php
@@ -50,8 +50,6 @@ function update1008to1009()
     $DELFROMDISPLAYPREF = [];
     $update_dir = __DIR__ . '/update_10.0.8_to_10.0.9/';
 
-    //TRANS: %s is the number of new version
-    $migration->displayTitle(sprintf(__('Update to %s'), '10.0.9'));
     $migration->setVersion('10.0.9');
 
     $update_scripts = scandir($update_dir);