+++ install/migrations/update_9.1.0_to_9.1.1.php
@@ -49,32 +49,7 @@ function update910to911()
     $updateresult     = true;
     $ADDTODISPLAYPREF = [];
 
-   //TRANS: %s is the number of new version
-    $migration->displayTitle(sprintf(__('Update to %s'), '9.1.1'));
     $migration->setVersion('9.1.1');
-
-    $backup_tables = false;
-   // table already exist but deleted during the migration
-   // not table created during the migration
-   // not table created during the migration
-    $newtables     = [];
-
-    foreach ($newtables as $new_table) {
-       // rename new tables if exists ?
-        if ($DB->tableExists($new_table)) {
-            $migration->dropTable("backup_$new_table");
-            $migration->displayWarning("$new_table table already exists. " .
-                                    "A backup have been done to backup_$new_table.");
-            $backup_tables = true;
-            $migration->renameTable("$new_table", "backup_$new_table");
-        }
-    }
-    if ($backup_tables) {
-        $migration->displayWarning(
-            "You can delete backup tables if you have no need of them.",
-            true
-        );
-    }
 
    // rectify missing right in 9.1 update
     if (countElementsInTable("glpi_profilerights", ['name' => 'license']) == 0) {