+++ install/migrations/update_9.2.1_to_9.2.2.php
@@ -53,8 +53,6 @@ function update921to922()
     $updateresult     = true;
     $ADDTODISPLAYPREF = [];
 
-   //TRANS: %s is the number of new version
-    $migration->displayTitle(sprintf(__('Update to %s'), '9.2.2'));
     $migration->setVersion('9.2.2');
 
     $migration->addConfig([