+++ install/migrations/update_9.5.x_to_10.0.0/recurrentchange.php
@@ -37,8 +37,6 @@
  * @var \Migration $migration
  */
 
-$migration->displayMessage("Adding recurrent changes");
-
 $default_charset = DBConnection::getDefaultCharset();
 $default_collation = DBConnection::getDefaultCollation();
 $default_key_sign = DBConnection::getDefaultPrimaryKeySignOption();