+++ src/Glpi/Controller/InstallController.php
@@ -34,32 +34,42 @@
 
 namespace Glpi\Controller;
 
+use Config;
+use DB;
+use Glpi\Cache\CacheManager;
 use Glpi\Exception\Http\AccessDeniedHttpException;
 use Glpi\Http\Firewall;
 use Glpi\Progress\ProgressStorage;
 use Glpi\Progress\StoredProgressIndicator;
 use Glpi\Security\Attribute\SecurityStrategy;
+use Migration;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\Routing\Attribute\Route;
 use Toolbox;
+use Update;
 
 class InstallController extends AbstractController
 {
     public const PROGRESS_KEY_INIT_DATABASE = 'init_database';
+    public const PROGRESS_KEY_UPDATE_DATABASE = 'update_database';
 
     public function __construct(
         private readonly ProgressStorage $progress_storage,
+        private readonly LoggerInterface $logger
     ) {
     }
 
-    #[Route("/install/init_database", methods: 'POST')]
+    #[Route("/Install/InitDatabase", methods: 'POST')]
     #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]
-    public function start_inserts(): Response
+    public function initDatabase(): Response
     {
         if (!isset($_SESSION['can_process_install'])) {
             throw new AccessDeniedHttpException();
         }
+
+        ini_set('max_execution_time', '300'); // Allow up to 5 minutes to prevent unexpected timeout
 
         $progress_indicator = new StoredProgressIndicator($this->progress_storage, self::PROGRESS_KEY_INIT_DATABASE);
 
@@ -70,6 +80,55 @@ class InstallController extends AbstractController
                 $progress_indicator->fail();
                 // Try to remove the config file, to be able to restart the process.
                 @unlink(GLPI_CONFIG_DIR . '/config_db.php');
+            }
+        });
+    }
+
+    #[Route("/Install/UpdateDatabase", methods: 'POST')]
+    #[SecurityStrategy(Firewall::STRATEGY_NO_CHECK)]
+    public function updateDatabase(): Response
+    {
+        if (!isset($_SESSION['can_process_update'])) {
+            throw new AccessDeniedHttpException();
+        }
+
+        ini_set('max_execution_time', '300'); // Allow up to 5 minutes to prevent unexpected timeout
+
+        if (!file_exists(GLPI_CONFIG_DIR . '/config_db.php')) {
+            throw new \RuntimeException('Missing database configuration file.');
+        } else {
+            include_once(GLPI_CONFIG_DIR . '/config_db.php');
+            if (!\class_exists(DB::class)) {
+                throw new \RuntimeException('Invalid database configuration file.');
+            }
+        }
+
+        /** @var \DBmysql $DB */
+        global $DB;
+        $DB = new DB();
+        $DB->disableTableCaching(); // Prevents issues on fieldExists upgrading from old versions
+
+        // Required, at least, by usage of `Plugin::unactivateAll()`
+        // FIXME: We should not have to load the configuration before running the update process.
+        Config::loadLegacyConfiguration();
+
+        $progress_indicator = new StoredProgressIndicator($this->progress_storage, self::PROGRESS_KEY_UPDATE_DATABASE);
+
+        $update = new Update($DB);
+        $update->setMigration(new Migration(GLPI_VERSION, $progress_indicator));
+        $update->setLogger($this->logger);
+
+        return new StreamedResponse(function () use ($update, $progress_indicator) {
+            try {
+                $update->doUpdates(
+                    current_version: $update->getCurrents()['version'],
+                    progress_indicator: $progress_indicator
+                );
+
+                // Force cache cleaning to ensure it will not contain stale data
+                (new CacheManager())->resetAllCaches();
+            } catch (\Throwable $e) {
+                $progress_indicator->fail();
             }
         });
     }